Changed gravity to a float

+++ migrations/20160801140929_batch.js
@@ -9,7 +9,7 @@ exports.up = function(knex, Promise) {
     table.dateTime('end_date');
     table.integer('expected_yield');
     table.integer('actual_yield');
-    table.integer('gravity');
+    table.float('gravity');
   })
 };
